requirements.txt
python-socketio
pyyaml
joblib
requests
